tweaked scopedServices test
