+++ grails-app/controllers/demo/AnotherDemoController.groovy
@@ -1,0 +1,6 @@
+package demo
+
+class AnotherDemoController {
+def anotherDemoService
+    def index() { }
+}
