+++ grails-app/controllers/demo/DemoController.groovy
@@ -1,0 +1,7 @@
+package demo
+
+class DemoController {
+def demoService
+
+    def index() { }
+}
