+++ test/unit/demo/AnotherDemoControllerSpec.groovy
@@ -1,0 +1,20 @@
+package demo
+
+import grails.test.mixin.TestFor
+import spock.lang.Specification
+
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(AnotherDemoController)
+class AnotherDemoControllerSpec extends Specification {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+    }
+}
